requirements.txt
Flask==3.0.0
PyMuPDF
python-docx==1.1.0
pandas==2.1.4
Werkzeug==3.0.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
